Add files via upload
This dataset represents the monthly inflation rates for all items in the U.S. from January 2019 to 2024. Inflation is measured by the Consumer Price Index (CPI) for all urban consumers, which reflects the average change in prices paid by consumers for a basket of goods and services over time.

The data is collected and published by the U.S. Bureau of Labor Statistics. The inflation rate in the dataset is expressed as a percentage, representing the year-over-year change in prices each month. This dataset is commonly used for economic analysis, forecasting inflation trends, and understanding the impact of inflation on various sectors, including supply chains and consumer behavior.